requirement.txt
pdfplumber
pymupdf
pillow
opencv-python
numpy
pandas
streamlit
openai
python-dotenv